+++ themes/dark.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from .balanced import *  # noqa: F401,F403